EOInterface/* numberOfRowsInTableView returns NSInteger now EOModeler/EOModelerDocument.m: do not insert a model into EditingContext. EOModeler/EOModelerEditor.m: avoid crash
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@30245 72102866-910b-0410-8b05-ffd578937521

+++ EOInterface/EOColumnAssociation.h
@@ -164,7 +164,7 @@ textShouldBeginEditing: (NSText *)fieldEditor;
 - (EOColumnAssociation *)editingAssociation;
 
 /* Providing table view data source.  */
-- (int)numberOfRowsInTableView: (NSTableView *)tableView;
+- (NSInteger)numberOfRowsInTableView: (NSTableView *)tableView;
 
 - (void)tableView: (NSTableView *)tableView
    setObjectValue: (id)object
